backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile backend/requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
click==8.1.7
    # via uvicorn
fastapi==0.104.1
    # via -r backend/requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
pydantic==2.5.1
    # via fastapi
pydantic-core==2.14.3
    # via pydantic
pyyaml==6.0.1
    # via -r backend/requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.24.0.post1
    # via -r backend/requirements.in
